Bug 1670874 - [mozproxy]Update mitm file structure r=tarek,perftest-reviewers
Differential Revision: https://phabricator.services.mozilla.com/D93330
