+++ testing/mozbase/mozproxy/mozproxy/server.py
@@ -3,7 +3,8 @@
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 from __future__ import absolute_import
-from mozproxy.backends.mitm import MitmproxyDesktop, MitmproxyAndroid
+from mozproxy.backends.mitm.android import MitmproxyAndroid
+from mozproxy.backends.mitm.desktop import MitmproxyDesktop
 
 _BACKENDS = {"mitmproxy": MitmproxyDesktop, "mitmproxy-android": MitmproxyAndroid}
